Allow to specify "max_time" explicitly
In some case the computed "max_time" does not fits our needs. Specifying it
explicitly is a simple way to fix the issue.

For example, if a benchmark have a variant running very quickly and one variant
running very slow. We want to keep the advantage of the default scheduling for
the fast one (with a small `sample_time`), without considering the slow variant
as "too_slow" too early.  Setting and higher "max_time" in this case will help
keeping