Minor clarification of the max_runtime logic
The max_time variable contained a "date", not a "duration". To clarify this, a
new variable is introduced: "deadline" contains the target date and max_time now
contains the max runtime for the variant.

This will be useful for a future changeset adding more way to control the target
duration.